% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{bwAverage4GRange}
\alias{bwAverage4GRange}
\title{Apply bigWigAverageOverBed to a GRange object}
\usage{
bwAverage4GRange(grange, bw.file, is.minMax = T)
}
\arguments{
\item{is.minMax}{}
}
\value{

}
\description{
Apply bigWigAverageOverBed to a GRange object
}
