% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/footprinting.R
\name{computeCentipedeMatrix}
\alias{computeCentipedeMatrix}
\title{Computes CENTIPEDE insertions matrix around given motif sites}
\usage{
computeCentipedeMatrix(insertions.gr, motifs, motif.ext = 100)
}
\arguments{
\item{motif.ext}{}
}
\value{

}
\description{
Computes CENTIPEDE insertions matrix around given motif sites
}
