% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{computeFIP}
\alias{computeFIP}
\title{Compute fraction of insertions in peaks}
\usage{
computeFIP(insertions.gr, peaks)
}
\arguments{
\item{peaks}{}
}
\value{

}
\description{
Compute fraction of insertions in peaks
}
