% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{computeFragmentProp}
\alias{computeFragmentProp}
\title{Function to compute the proportion of fragments within a given range
of fragment lengths.}
\usage{
computeFragmentProp(frag.lengths, min.length, max.length)
}
\arguments{
\item{max.length}{}
}
\value{

}
\description{
Function to compute the proportion of fragments within a given range
of fragment lengths.
}
