% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/footprinting.R
\name{computeInsertionProfile}
\alias{computeInsertionProfile}
\title{Compute insertions meta-profile centered on a given set of regions}
\usage{
computeInsertionProfile(insertions.gr, regions, strand.spec = F)
}
\arguments{
\item{regions}{}
}
\value{

}
\description{
Compute insertions meta-profile centered on a given set of regions
}
