% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{computeProfile}
\alias{computeProfile}
\title{Computes a metaprofile given GRange object and *.bw file}
\usage{
computeProfile(region.gr, bw.file, region.names = NULL)
}
\arguments{
\item{region.names}{}
}
\value{

}
\description{
Computes a metaprofile given GRange object and *.bw file
}
