% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/footprinting.R
\name{computeRocStats}
\alias{computeRocStats}
\title{Compute ROC statistics for TF footprinting vs ChIP-seq comparison}
\usage{
computeRocStats(prob, labels, name = "")
}
\arguments{
\item{name}{}
}
\value{

}
\description{
Compute ROC statistics for TF footprinting vs ChIP-seq comparison
}
