% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/footprinting.R
\name{computeTn5Bias}
\alias{computeTn5Bias}
\title{Compute Tn5 sequence preference from a subset of insertions}
\usage{
computeTn5Bias(insertions.gr, bs.genome, ext = 10, chrs = c("chr16",
  "chr19"))
}
\arguments{
\item{chrs}{}
}
\value{

}
\description{
Compute Tn5 sequence preference from a subset of insertions
}
