% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{computeTssEnrichmentScore}
\alias{computeTssEnrichmentScore}
\title{Computes TSS enrichment score}
\usage{
computeTssEnrichmentScore(insertions.gr, txdb, up = 2000, down = 2000,
  win.size = 50)
}
\arguments{
\item{win.size}{}
}
\value{

}
\description{
Computes TSS enrichment score
}
