% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/footprinting.R
\name{computeVPlotMatrix}
\alias{computeVPlotMatrix}
\title{Compute V-plot insertion matrix centered on a given set of regions}
\usage{
computeVPlotMatrix(insertions.gr, regions)
}
\arguments{
\item{regions}{}
}
\value{

}
\description{
Compute V-plot insertion matrix centered on a given set of regions
}
