% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{getSummit}
\alias{getSummit}
\title{Get summit position for GenomicRanges}
\usage{
getSummit(regions.gr)
}
\arguments{
\item{regions.gr}{}
}
\value{

}
\description{
Get summit position for GenomicRanges
}
