% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{listFiles}
\alias{listFiles}
\title{Better version of listFiles}
\usage{
listFiles(path, is.file = T, pattern = NULL, maxdepth = NULL)
}
\arguments{
\item{maxdepth}{}
}
\value{

}
\description{
Better version of listFiles
}
