% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactions.R
\name{loadInteractionDb}
\alias{loadInteractionDb}
\title{Load interaction database (4DGenome)}
\usage{
loadInteractionDb(genome = "mm10")
}
\arguments{
\item{genome}{}
}
\value{

}
\description{
Load interaction database (4DGenome)
}
