% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{plotFragmentLengthDist}
\alias{plotFragmentLengthDist}
\title{Plot fragment length distribution for ATAC-seq PE data.}
\usage{
plotFragmentLengthDist(insertions.gr, min.length = 0, max.length = 500)
}
\arguments{
\item{max.length}{}
}
\value{

}
\description{
Plot fragment length distribution for ATAC-seq PE data.
}
