% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/footprinting.R
\name{plotInsertionProfile}
\alias{plotInsertionProfile}
\title{Plots insertion meta-profile around a motif given a insertion profile object}
\usage{
plotInsertionProfile(ins.profile, motif.w, col = "black")
}
\arguments{
\item{col}{}
}
\value{

}
\description{
Plots insertion meta-profile around a motif given a insertion profile object
}
