% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{plotTssEnrichment}
\alias{plotTssEnrichment}
\title{Plots Tn5 insertion metaprofile around TSS and compute enrichment score}
\usage{
plotTssEnrichment(insertions.gr, txdb, up = 2000, down = 2000,
  win.size = 50)
}
\arguments{
\item{win.size}{}
}
\value{

}
\description{
Plots Tn5 insertion metaprofile around TSS and compute enrichment score
}
