% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/footprinting.R
\name{plotVPlot}
\alias{plotVPlot}
\title{Plots V-Plot for a before compute V-Plot data object}
\usage{
plotVPlot(vplot.data, motif.w, smooth.w = 5)
}
\arguments{
\item{motif.w}{}
}
\value{

}
\description{
Plots V-Plot for a before compute V-Plot data object
}
