% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactions.R
\name{prepareInteractionDb}
\alias{prepareInteractionDb}
\title{Load interactions from *.bed files and save as RData file}
\usage{
prepareInteractionDb(interA.file, interB.file, inter.meta, out.rdata)
}
\arguments{
\item{inter.meta}{}
}
\value{

}
\description{
Load interactions from *.bed files and save as RData file
}
