% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{readInsertions}
\alias{readInsertions}
\title{Read Tn5 Insertions from paired-end bam file}
\usage{
readInsertions(bam.file, genome, is.tn5Adjusted = F, mapq = 30)
}
\arguments{
\item{mapq}{}
}
\value{

}
\description{
Read Tn5 Insertions from paired-end bam file
}
