% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{readInsertionsFromSE}
\alias{readInsertionsFromSE}
\title{Read Tn5 Insertions from single-end bam file}
\usage{
readInsertionsFromSE(bam.file, genome, is.tn5Adjusted = F, mapq = 30)
}
\arguments{
\item{mapq}{}
}
\value{

}
\description{
Read Tn5 Insertions from single-end bam file
}
