% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{savePlots}
\alias{savePlots}
\title{Plot ggplot to multiple files defined by figure suffix}
\usage{
savePlots(gg.plot, file.name, figure.suffix = c("pdf", "eps"),
  base_height = 4, base_width = NULL)
}
\arguments{
\item{base_width}{}
}
\value{

}
\description{
Plot ggplot to multiple files defined by figure suffix
}
