% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactions.R
\name{subsetInteractionsByGene}
\alias{subsetInteractionsByGene}
\title{Subset interactions by a given gene target}
\usage{
subsetInteractionsByGene(interactions, gene)
}
\arguments{
\item{gene}{}
}
\value{

}
\description{
Subset interactions by a given gene target
}
