% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactions.R
\name{subsetInteractionsByRegion}
\alias{subsetInteractionsByRegion}
\title{Subsets interactions by a given regions
(--> get interactions within region)}
\usage{
subsetInteractionsByRegion(interactions, region)
}
\arguments{
\item{region}{}
}
\value{

}
\description{
Subsets interactions by a given regions
(--> get interactions within region)
}
